+++ src/auth/authClient.ts
@@ -1,11 +1,13 @@
-import { DefaultTransporter } from '../transporters';
+import { DefaultTransporter, IRequestOptions } from '../transporters';
 
 /**
  * 抽象認証クライアント
  */
 export abstract class AuthClient {
-    public abstract async fetch(url: string, options: RequestInit, expectedStatusCodes: number[]): Promise<Response>;
-    public abstract async getAccessToken(): Promise<string>;
+    public abstract fetch(
+        url: string, options: RequestInit, requestOptions: IRequestOptions, expectedStatusCodes: number[]
+    ): Promise<Response>;
+    public abstract getAccessToken(): Promise<string>;
 }
 /**
  * テスト認証クライアント
@@ -14,8 +16,10 @@ export abstract class AuthClient {
 /* istanbul ignore next */
 export class StubAuthClient implements AuthClient {
     // tslint:disable-next-line:prefer-function-over-method
-    public async fetch(url: string, options: RequestInit, expectedStatusCodes: number[]): Promise<Response> {
-        return (new DefaultTransporter(expectedStatusCodes)).fetch(url, options);
+    public async fetch(
+        url: string, options: RequestInit, requestOptions: IRequestOptions, expectedStatusCodes: number[]
+    ): Promise<Response> {
+        return (new DefaultTransporter(expectedStatusCodes)).fetch(url, options, requestOptions);
     }
     // tslint:disable-next-line:prefer-function-over-method
     public async getAccessToken(): Promise<string> {