+++ src/transporter.spec.ts
@@ -48,7 +48,8 @@ describe('fetch()', () => {
                 .get('/uri')
                 .reply(statusCode, body);
 
-            const result = await transporter.fetch(`${API_ENDPOINT}/uri`, {}).then(async (res) => res.json());
+            const result = await transporter.fetch(`${API_ENDPOINT}/uri`, {}, {})
+                .then(async (res) => res.json());
             assert.deepEqual(result, body);
             sandbox.verify();
             assert(scope.isDone());
@@ -79,7 +80,8 @@ describe('fetch()', () => {
                 .get('/uri')
                 .reply(statusCode, body);
 
-            const result = await transporter.fetch(`${API_ENDPOINT}/uri`, {}).catch((err) => err);
+            const result = await transporter.fetch(`${API_ENDPOINT}/uri`, {}, {})
+                .catch((err) => err);
 
             assert(result instanceof Error);
             assert.equal((<RequestError>result).code, statusCode);
@@ -87,6 +89,23 @@ describe('fetch()', () => {
             sandbox.verify();
             assert(scope.isDone());
         });
+    });
+
+    it('timeoutを設定してもレスポンスを取得できるはず', async () => {
+        const body: any = { key: 'value' };
+
+        const transporter = new DefaultTransporter([OK]);
+
+        scope = nock(API_ENDPOINT)
+            .get('/uri')
+            .reply(OK, body);
+
+        const result = await transporter
+            .fetch(`${API_ENDPOINT}/uri`, {}, { timeout: 10000 })
+            .then(async (res) => res.json());
+        assert.deepEqual(result, body);
+        sandbox.verify();
+        assert(scope.isDone());
     });
 
     it('レスポンスボディがjsonでなければ、適切なエラーが投げられるはず', async () => {
@@ -99,7 +118,8 @@ describe('fetch()', () => {
             .get('/uri')
             .reply(statusCode, body);
 
-        const result = await transporter.fetch(`${API_ENDPOINT}/uri`, {}).catch((err) => err);
+        const result = await transporter.fetch(`${API_ENDPOINT}/uri`, {}, {})
+            .catch((err) => err);
         assert(result instanceof Error);
         assert.equal((<RequestError>result).code, statusCode);
         assert.equal((<RequestError>result).message, body);